Moved components to frontend root
