+++ src/frontend/componenets/Header.tsx
@@ -1,15 +1,15 @@
 // @ts-ignore
-import classes from "../frontend/styles/components/Header.module.scss";
+import classes from "../styles/components/Header.module.scss";
 
 import React from "react";
 import {useRouter} from "next/router";
 import Link from "next/link";
 import clsx from "clsx";
 import Button from "@mui/material/Button";
-import ThemeSwitch from "../frontend/fragments/ThemeSwitch.Header";
-import { useAccount } from "../pages/_app";
-import { connected, connectWallet, network } from "../utils/frontend/metamask";
-import { getAccountDisplayValue } from "../utils/frontend/helpers";
+import ThemeSwitch from "../fragments/ThemeSwitch.Header";
+import { useAccount } from "../../pages/_app";
+import { connected, connectWallet, network } from "../../utils/frontend/metamask";
+import { getAccountDisplayValue } from "../../utils/frontend/helpers";
 
 type Route =
 {
